nx_vms_server module warnings fixes
<==== Mercurial data ====>
Branch: vms_4.0_server

+++ vms/server/nx_vms_server/src/network/universal_tcp_listener.cpp
@@ -179,7 +179,7 @@ void QnUniversalTcpListener::updateCloudConnectState(
                 std::move(registrationOnMediatorRetryPolicy));
         cloudServerSocket->listen(0);
         m_multipleServerSocket->addSocket(std::move(cloudServerSocket));
-        m_cloudSocketIndex = m_multipleServerSocket->count() - 1;
+        m_cloudSocketIndex = (int) m_multipleServerSocket->count() - 1;
     }
     else
     {
